refactor(draug): A.4a — move task_store + draug.rs to daemon
Continues the file-by-file migration. Compositor still builds clean
(same 93 warnings as baseline) thanks to thin re-export shims that
keep all existing import paths resolving.

* draug-daemon/src/task_store.rs (369 LoC) — Synapse VFS-backed
  refactor-task store. Moved verbatim, with `compositor::refactor_types`
  → `crate::refactor_types` import.
* draug-daemon/src/draug.rs (1844 LoC) — DraugDaemon struct, friction
  sensor (FRICTION_QUICK_CLOSE / RAGE_CLICK / BACKSPACE_SPAM
  constants), hibernation gate, plan-mode state machine, task levels,
  WASM crash recording, status persistence. Moved verbatim. Inner
  `crate::refactor_types::*` references stay valid because that
  module already lives in the same crate (committed earlier).
* draug-daemon/Cargo.toml: enabled libfolk's `mcp` feature so the
  Draug analyse_telemetry path can call `libfolk::mcp::client`.
* Compositor's draug.rs and mcp_handler/task_store.rs reduced to
  one-line `pub use draug_daemon::draug::*;` shims.

Verified: cargo check + cargo check --bin compositor pass clean
for both crates. No new warnings.
