+++ userspace/draug-daemon/src/lib.rs
@@ -20,3 +20,5 @@
 extern crate alloc;
 
 pub mod refactor_types;
+pub mod task_store;
+pub mod draug;
